-l and -c

+++ set_snippets.sh
@@ -1,5 +1,38 @@
 #!/bin/bash
 # elpa/elpy-20220627.1416/
+
+#If /set_snippets.sh -c is called nothing happens, just checking
+
+
+CHECKONLY=1
+
+if [ "$#" -eq  "0" ] ;  then
+#    echo "No arguments supplied"
+    CHECKONLY=1
+else
+    if [ ! -z "$1" ] ; then
+	if [ "$1" == "-c" ] ; then
+#	    echo "P1 is -c"
+	    CHECKONLY=0
+	    echo "ReadOnly Mode"
+	else	    
+#	    echo "P1 is not -c"
+#	    echo "P1 is {$1}"
+	    CHECKONLY=1
+	fi
+    else
+#	echo "P1 is not set"
+	CHECKONLY=1
+    fi    
+fi
+
+#echo ${CHECKONLY}
+#echo "$1"
+
+if [ "$1" = "-l" ] ; then
+    echo "Delete Mode"
+fi
+
 FILPA=$(pwd)
 cd elpa
 ELPY=$(ls -d elpy*)
@@ -12,40 +45,48 @@ for file in snippets/*; do
 	OLDFILE=${FULLFILE}_old
 	if [ -L ${FULLFILE} ] ; then
 		if [ -e ${FULLFILE} ] ; then
-      			echo "Good link: ${FULLFILE} "
+      		    echo "Good link: ${FULLFILE} "
+		    #Check option -c
+		    if [ "$CHECKONLY" = 1 ] ; then
 			if [ -f ${OLDFILE} ]; then
 				echo ${OLDFILE}
 				rm ${OLDFILE}
 			fi
-			mv ${FULLFILE} ${OLDFILE}
-
+			if [ "$1" = "-l" ] ; then
+			    echo "Remove set removing: ${FULLFILE} "
+			    rm ${FULLFILE}
+			else
+			    echo "Moving to old_file instead of deleting"
+			    mv ${FULLFILE} ${OLDFILE}			
+			fi
+		    fi
+		    
 		else
-      			echo "Broken link: ${FULLFILE} "
+      		    echo "Broken link: ${FULLFILE} "
+		    if [ "$CHECKONLY" = 1 ] ; then
+			echo "rm ${FULLFILE}"
 			rm ${FULLFILE}
+		    fi		    
    		fi
 	elif [ -e ${FULLFILE} ] ; then
-		echo "Not a link: ${FULLFILE}"
+	    echo "Not a link: ${FULLFILE}"
+	    if [ "$CHECKONLY" = 1 ] ; then
 		if [ -f ${OLDFILE} ]; then
 			echo ${OLDFILE}
 			rm ${OLDFILE}
 		fi
+		echo "Moving as it's not a link"
                 mv ${FULLFILE} ${OLDFILE}
+	    fi	    
 	else
    		echo "Missing: ${FULLFILE}"
 	fi
-
-	ln -s ${ORIGFILE} ${FULLFILE}
+	
+	if [ "$CHECKONLY" = 1 ] ; then
+	    echo "Setting Link"
+	    ln -s ${ORIGFILE} ${FULLFILE}
+	fi	
 done
 
-#	if [ -L ${FULLFILE} ] && [ -e ${FULLFILE} ]; then
-#		OLDFILE=${FULLFILE}_old		
-#		if [ -L ${OLDFILE} ] && [ -e ${OLDFILE} ]; then
-#			echo ${OLDFILE}
-#			rm ${OLDFILE}
-#		fi
-#		echo ${FULLFILE}
-#		mv ${FULLFILE} ${OLDFILE}
-#	else
-#	       	ln -s ${ORIGFILE} ${FULLFILE}; 
-#	fi
-#done
+
+
